refactor: renames main module
